Move navigator to components
