+++ src/lib/pages/workflow-history-json.svelte
@@ -2,7 +2,7 @@
   import { page } from '$app/stores';
   import ToggleSwitch from '$lib/holocene/toggle-switch.svelte';
 
-  import WorkflowJsonNavigator from '$lib/holocene/workflow-json-navigator.svelte';
+  import WorkflowJsonNavigator from '$lib/components/workflow/workflow-json-navigator.svelte';
   import { fetchAllEvents, fetchRawEvents } from '$lib/services/events-service';
   import { authUser } from '$lib/stores/auth-user';
   import { decodeURIForSvelte } from '$lib/utilities/encode-uri';
