+++ src/routes/(app)/import/events/[namespace]/[workflow]/[run]/history/json/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import WorkflowJsonNavigator from '$lib/holocene/workflow-json-navigator.svelte';
+  import WorkflowJsonNavigator from '$lib/components/workflow/workflow-json-navigator.svelte';
 
   import { importEvents } from '$lib/stores/import-events';
 </script>
